pyproject.toml
[project]
name = "test-service-at-hotel-level"
version = "0.1.0"
description = ""
authors = [
    {name = "Alex X",email = "dowsha@yandex.ru"}
]
readme = "README.md"
requires-python = ">=3.9.6"
dependencies = [
    "django (==4.2)",
    "djangorestframework (==3.14.0)",
    "ruff (==0.3.5)",
    "pre-commit (==3.6.2)",
    "pytest (==7.4.4)",
    "pytest-django (==4.8.0)",
    "drf-yasg (==1.20.0)",
    "setuptools (>=78.1.0,<79.0.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry]
packages = [{ include = "test_service_at_hotel_level" }]
